Fixed binary string functions in GM8
GM8 apparently uses a null-pointer internally for the empty string, and translates that to a pointer to a constant 0-length 0-terminated string when calling extension functions. However, that constant string doesn't have a length prefix, so attempting to read it will return some bytes from the function it's defined in. This doesn't appear to happen in GM8.1.